rm redundancy
